+++ src/Firebase/DynamicLink/ShortenLongDynamicLink/ApiRequest.php
@@ -22,7 +22,7 @@ final class ApiRequest implements RequestInterface
 
         $headers = [
             'Content-Type' => 'application/json; charset=UTF-8',
-            'Content-Length' => \mb_strlen((string) $body),
+            'Content-Length' => $body->getSize(),
         ];
 
         $this->wrappedRequest = new Request('POST', $uri, $headers, $body);