+++ src/Firebase/Messaging/Http/Request/SendMessage.php
@@ -21,7 +21,7 @@ final class SendMessage implements RequestInterface
         $body = stream_for(\json_encode(['message' => $message]));
         $headers = [
             'Content-Type' => 'application/json; charset=UTF-8',
-            'Content-Length' => \mb_strlen((string) $body),
+            'Content-Length' => $body->getSize(),
         ];
 
         $this->wrappedRequest = new Request('POST', $uri, $headers, $body);